misc warnings and bugs

+++ src/filter.cpp
@@ -123,7 +123,7 @@ FilterItem::FilterItem(const QString& filterPattern, bool caseSensitive)
       // if a hyphen was specified, but no max value after it, it means
       // all values above min
       if (levelString.isEmpty())
-        maxLevel = INT_MAX;
+        maxLevel = UINT8_MAX;
       else
       {
         // get the max level
